+++ source/Application/Controller/Admin/OrderOverview.php
@@ -69,7 +69,6 @@ class OrderOverview extends \OxidEsales\Eshop\Application\Controller\Admin\Admin
 
     /**
      * Returns user payment used for current order. In case current order was executed using
-     * credit card and user payment info is not stored in db (if \OxidEsales\Eshop\Core\Config::blStoreCreditCardInfo = false),
      * just for preview user payment is set from oxPayment
      *
      * @param object $oOrder Order object